Add system role

+++ src/main/kotlin/com/liah/doribottle/domain/user/Admin.kt
@@ -1,6 +1,9 @@
 package com.liah.doribottle.domain.user
 
+import com.liah.doribottle.common.error.exception.BusinessException
+import com.liah.doribottle.common.error.exception.ErrorCode
 import com.liah.doribottle.domain.common.SoftDeleteEntity
+import com.liah.doribottle.extension.isSystem
 import com.liah.doribottle.service.user.dto.AdminDto
 import com.liah.doribottle.service.user.dto.AdminSimpleDto
 import jakarta.persistence.*
@@ -52,6 +55,9 @@ class Admin(
         protected set
 
     override fun delete() {
+        if (this.isSystem())
+            throw BusinessException(ErrorCode.SYSTEM_DELETE_NOT_ALLOWED)
+
         this.loginId = "Deleted ${UUID.randomUUID()}"
         super.delete()
     }
